Move Tagline constant to version package
Refactors the Tagline constant from root package to a dedicated
version package for better organization and separation of concerns.

Changes:
- Create new version/version.go with Tagline constant
- Update main.go to import and use version.Tagline
- Update ui/model.go to import and use version.Tagline
- Improve UI layout by separating title from tagline
- Remove old version.go from root
