+++ version/version.go
@@ -1,4 +1,4 @@
-package main
+package version
 
 // Tagline is the application's tagline used in help text and documentation
 const Tagline = "I'm Rocha, and I manage coding agents"
